Move algebra library into a distinct sub-project
