+++ frontends/algebra/src/main/stainless/algebra/AbelianGroup.scala
