+++ frontends/algebra/src/main/stainless/algebra/BoundedJoinSemilattice.scala
