+++ frontends/algebra/src/main/stainless/algebra/BoundedMeetSemilattice.scala
