+++ frontends/algebra/src/main/stainless/algebra/Eq.scala
