+++ frontends/algebra/src/main/stainless/algebra/Group.scala
