+++ frontends/algebra/src/main/stainless/algebra/JoinSemilattice.scala
