+++ frontends/algebra/src/main/stainless/algebra/MeetSemilattice.scala
