+++ frontends/algebra/src/main/stainless/algebra/Monoid.scala
