+++ frontends/algebra/src/main/stainless/algebra/PartialOrder.scala
