+++ frontends/algebra/src/main/stainless/algebra/Ring.scala
