+++ frontends/algebra/src/main/stainless/algebra/Semigroup.scala
