+++ frontends/algebra/src/main/stainless/algebra/TotalOrder.scala
